cloud-run-deployment/adk-agent/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
google-adk>=1.0.0
python-dotenv>=1.0.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
